+++ backend/src/main/java/com/example/backend/dao/PaymentDAO.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PaymentDAO extends JpaRepository<Payment, Long> {
+public interface PaymentDAO extends JpaRepository<Payment, Integer> {
     List<Payment> findByUserIdAndIsActive(int UserId, Boolean isActive);
 }
